refactor(oxfmt): Misc internal refactorings (#16726)
- Move `init_rayon` from `cli/command.rs` to `cli/init.rs`
- Move config related functions from `cli/format.rs` to `core/config.rs`
- Create `BufWrite` inside of `cli/format.rs`

Purely a refactoring to move and organize the code.
I'll merge it after CI completes.
