+++ apps/oxfmt/src/cli/mod.rs
@@ -8,5 +8,5 @@ mod walk;
 
 pub use command::format_command;
 pub use format::FormatRunner;
-pub use init::{init_miette, init_tracing};
+pub use init::{init_miette, init_rayon, init_tracing};
 pub use result::CliRunResult;
